Project files
This is the finished "product". It's kind of what it was supposed to be.
Some problems that iv'e encountered were that i had problems with include:

The camera functionality (it is not made to have an camera atm because of some trouble with the namespaces.). 
The mouse coordinates function interprets the entire forms window, which was not the original aim of the project.

That's pretty much it, i will continue working on this. if you want to check up on the progress, you can be my guest.
That is all for now, hope this is worth something.